requirements.txt
# Company Research & Account Plan Agent Dependencies

# Core async functionality
asyncio-tools==0.1.2

# Web scraping and extraction
crawl4ai>=0.3.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
duckduckgo-search==4.1.0

# AI and LLM integrations
crewai==0.1.7
groq==0.4.2
mistralai==0.0.8

# Data processing
pydantic==2.5.3
python-dotenv==1.0.0
tqdm==4.66.1

# File handling (for Windows compatibility)
pywin32==306; sys_platform == 'win32'

# Additional utilities
pathlib2==2.3.7
